workarounds for Codon compatibility
Codon File reads currently output strings not bytes.
This is an issue as struct input is bytes.

Two different approaches are taken to address this.

For weights, a new 'fdfile' class is used that opens and reads using file descripters in python interpeter.
This approach is good for reading large amounts of binary data quickly.

For tokens, new functions pack the characters as unsigned bytes, then unpack as int or float.

To allow functionality in python and codon, a dummy 'python' module is create to allow importing struct, and a dummy '@python' decorator is added to allow the codon specification not to interfere with python.